// src/app/shared/utils/models/transformation-types.model.ts
// tslint:disable: variable-name
// tslint:disable: no-inferrable-types
// tslint:disable: no-use-before-declare
export class TransformationTypesModel {
    transformation_type_id: string;
    name: string;
    source_id: string;
    is_enabled: boolean = true;
    branch_id: number;
    mstatus: string;
    disabled: boolean;
    user_id: any;
}